requirements.txt
# this is the requirements.txt file...

#
# DEVELOPMENT / PRODUCTION
#

python-dotenv
pandas
requests


flask
gunicorn

#
# TEST
#

pytest
